# [X86] Optimize gru and softmax