### 优化内容
- 优化了gru和softmax两个op的计算
  - gru：
    - 一些std::vector改成引用类型
    - 去掉Tensor::Slice的调用，直接计算指针
  - softmax：
    - 当输入是2-D，且axis=1时，不对输入输出进行Resize

- develop
  ```
  commit 0699ee909e0868b000f99c47f71e77a690868e48
  Author: xiaogang <txg4794@163.com>
  Date:   Fri Feb 14 10:24:00 2020 +0800

      fix: move some necessary op from extra to basic, (#2872)
  ```
  - StepRNN的运行时间：**0.0211 ms**
  ```
  I0214 04:45:03.967396  6551 generate_program_pass.h:37] insts.size 42
  I0214 04:45:06.095240  6551 test_step_rnn_lite_x86.cc:90] warmup: 10, repeats: 100000, spend 0.021101 ms in average.
  ```
- PR #2877 
  - StepRNN的运行时间：**0.0186 ms**
  ```
  I0214 05:41:01.555439 21354 generate_program_pass.h:37] insts.size 42
  I0214 05:41:03.431866 21354 test_step_rnn_lite_x86.cc:90] warmup: 10, repeats: 100000, spend 0.0186351 ms in average.
  ```

### gperftools分析结果
- develop：
  - GRUCompute::Run，14.69%
    - TensorLite::Slice、TensorLite的构造、析构函数（临时Tensor的使用），加一起占了不少时间。
    - 还有一些std::vector和DDim的构造函数
  - SoftmaxCompute::Run，7.14%
    - softmax计算的时候，会将输入输出Tensor先Resize成一个2-D Tensor，使用到了临时Tensor，并且调用`ShareDataWith`和输入、输出Tensor共享数据。profile火山图中显示`ShareDataWith`占了很大的时间比。实际上，当输入、输出Tensor本身已经是2-D Tensor，则不需要这个Resize过程。
    - `SoftmaxFunctor`中存在std::vector的使用，经分析发现，其实没有必要使用std::vector。

![image](https://user-images.githubusercontent.com/12538138/74503909-73043680-4f2d-11ea-90d0-d9c0363e3ff5.png)

![image](https://user-images.githubusercontent.com/12538138/74504273-a1364600-4f2e-11ea-934d-faff3e6665a6.png)

- PR #2877 ：
  - GRUCompute::Run，11.8%
    - 因为gru中还有移除有临时Tensor的使用，所以仍有TensorLite的构造、析构时间，但占比明显少了很多。
  - SoftmaxCompute::Run，4.35%
    - 完全去除了所有TensorLite定义、ShareDataWith和std::vector时间占用。

![image](https://user-images.githubusercontent.com/12538138/74504960-dba0e280-4f30-11ea-9085-b53d8e126102.png)

![image](https://user-images.githubusercontent.com/12538138/74504985-e9566800-4f30-11ea-9c0e-574e8852df6e.png)